+++ packages/components/src/Accordion/AccordionContext.tsx
@@ -24,51 +24,34 @@
 
  */
 
-export * from './Accordion'
-export * from './ActionList'
-export * from './Avatar'
-export * from './Badge'
-export * from './Banner'
-export * from './Button'
-export * from './Calendar'
-export * from './Card'
-export * from './Chip'
-export * from './DateTimeFormat'
-export * from './Divider'
-export * from './Form'
-export * from './Icon'
-export * from './Layout'
-export * from './Link'
-export * from './List'
-export * from './Menu'
-export * from './Modal'
-export * from './PageSize'
-export * from './Pagination'
-export * from './Popover'
-export * from './Spinner'
-export * from './Sidebar'
-export * from './Table'
-export * from './Tabs'
-export * from './Tooltip'
-export * from './Text'
-export * from './VisuallyHidden'
+import { SpacingSizes } from '@looker/design-tokens'
+import { createContext } from 'react'
+import noop from 'lodash/noop'
+import { IndicatorIcons } from './Accordion'
 
-export * from './utils'
+export interface AccordionContextProps {
+  indicatorGap: SpacingSizes
+  indicatorIcons: IndicatorIcons
+  indicatorSize: SpacingSizes
+  indicatorPosition: 'left' | 'right'
+  isOpen: boolean
+  toggleOpen: (isOpen: boolean) => void
+  onClose?: () => void
+  onOpen?: () => void
+}
 
-export { ComponentsProvider } from '@looker/components-providers'
+export const accordionContextDefaults: AccordionContextProps = {
+  indicatorGap: 'xsmall',
+  indicatorIcons: {
+    close: 'CaretDown',
+    open: 'CaretUp',
+  },
+  indicatorPosition: 'right',
+  indicatorSize: 'large',
+  isOpen: false,
+  toggleOpen: (isOpen: boolean) => noop(isOpen),
+}
 
-/** Provided here for backwards compatibility.
- * @TODO - Remove before 1.0
- **/
-
-export {
-  GlobalStyle,
-  palette,
-  radii,
-  semanticColors,
-  SemanticColor,
-  SemanticColors,
-  SpacingSizes,
-  theme,
-  Theme,
-} from '@looker/design-tokens'
+export const AccordionContext = createContext<AccordionContextProps>({
+  ...accordionContextDefaults,
+})